+++ src/components/Onboarding/OnboardingStepWrapper.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const OnboardingStepWrapper = ({title, description, children}) => {
+  return (
+    <>
+      <h2>{title}</h2>
+      <p>{description}</p>
+      {children}
+    </>
+  )
+}
+
+export default OnboardingStepWrapper
